7주차/ddpm_project/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ddpm-research-artifact"
version = "0.2.0"
description = "DDPM research repo with Hydra + reproducible toy sanity pipeline"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "torch>=2.1",
  "torchvision>=0.16",
  "lightning>=2.2",
  "hydra-core>=1.3",
  "omegaconf>=2.3",
  "wandb>=0.16",
  "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.5",
  "black>=24.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [".git", ".venv", "outputs", "data"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["."]
